[ALPS06134707] media: camsys: fix stagger seamless switch
1. feature control value is large enough to use int64
after supporting general seamless switch, change to type
to int64 or it may use max value if value is too large.

2. in seamless switch, pipe usage, such as camsv pipe,
may vary between switching, check pipe usage by request
than ctx.

MTK-Commit-Id: dda90c0d7b22b0691c84e355700867f4c8b3ee2d

CR-Id: ALPS06134707
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/camsys/isp7_1/cam/mtk_cam-raw.h
@@ -254,9 +254,9 @@ struct mtk_raw_pipeline {
 	unsigned long enabled_dmas;
 	/* resource controls */
 	struct v4l2_ctrl_handler ctrl_handler;
-	int feature_pending;
-	int feature_pending_try;
-	int feature_active;
+	s64 feature_pending;
+	s64 feature_pending_try;
+	s64 feature_active;
 	bool enqueued_tg_flash_req; /* need a better way to collect the request */
 	struct mtk_cam_tg_flash_config tg_flash_config;
 	/* TODO: merge or integrate with mtk_cam_resource_config */
